+++ tests/unit/errors/test_api_errors.py
@@ -13,7 +13,7 @@
 #  limitations under the License.
 
 import pytest
-from argilla_server.errors.api_errors import APIErrorHandler
+from argilla_server.errors import APIErrorHandler
 from argilla_server.errors.base_errors import (
     EntityAlreadyExistsError,
     EntityNotFoundError,